requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==3.2.14
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-yasg
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements.in
faker==12.0.1
    # via mixer
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.1
    # via jinja2
mixer==7.2.2
    # via -r requirements.in
packaging==21.3
    # via drf-yasg
psycopg2-binary==2.9.3
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via faker
pytz==2022.1
    # via
    #   django
    #   djangorestframework
requests==2.28.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
typing-extensions==4.3.0
    # via
    #   asgiref
    #   faker
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.10
    # via requests
